fragment 전환 방식으로 수정
