+++ app/src/main/AndroidManifest.xml
@@ -4,6 +4,7 @@
     package="com.example.email">
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
