+++ app/src/main/java/com/example/email/App.kt
@@ -1,0 +1,10 @@
+package com.example.email
+
+import android.app.Application
+import androidx.lifecycle.MutableLiveData
+
+class App : Application() {
+    companion object{
+        val mailType  = MutableLiveData(0)
+    }
+}
