+++ app/src/main/java/com/example/email/fragments/SettingFragment.kt
@@ -7,10 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.email.R
 import com.example.email.activities.MainActivity
-import com.example.email.adapters.MailAdapter
 import com.example.email.databinding.FragmentSettingBinding
 import com.example.email.viewmodels.SettingViewModel
 
@@ -27,14 +25,13 @@ class SettingFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         binding = DataBindingUtil.inflate(inflater,R.layout.fragment_setting,container,false)
-        binding.lifecycleOwner = viewLifecycleOwner
-
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        binding.lifecycleOwner = viewLifecycleOwner
         binding.viewModel = viewModel
 
         viewModel.emailText.value = "Email : ${MainActivity.email}"
